+++ src/Dto/Order/Core/OrderPaymentDetailsDto.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace malpka32\InPostBuySdk\Dto\Order\Core;
 
 use malpka32\InPostBuySdk\Collection\OrderPaymentCollection;
+use malpka32\InPostBuySdk\Dto\Order\OrderPaymentType;
 
 /**
  * Szczegóły płatności zamówienia.
@@ -12,7 +13,7 @@ use malpka32\InPostBuySdk\Collection\OrderPaymentCollection;
 final class OrderPaymentDetailsDto
 {
     public function __construct(
-        public ?string $selectedPaymentType = null,
+        public OrderPaymentType|string|null $selectedPaymentType = null,
         public ?OrderPaymentCollection $payments = null,
     ) {
     }