+++ src/Dto/Order/Core/OrderPaymentDto.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace malpka32\InPostBuySdk\Dto\Order\Core;
 
+use malpka32\InPostBuySdk\Dto\Order\OrderPaymentType;
+
 /**
  * Pojedyncza płatność przypisana do zamówienia.
  */
@@ -11,7 +13,7 @@ final class OrderPaymentDto
 {
     public function __construct(
         public ?string $paymentId = null,
-        public ?string $paymentType = null,
+        public OrderPaymentType|string|null $paymentType = null,
         public ?\DateTimeInterface $paymentDate = null,
     ) {
     }